refactor(components): Rename new-lia-card to lia-card and refactor depended components
