+++ app/src/app/components/lia-card/index.ts
@@ -1,0 +1,1 @@
+export { LIA_CARD_DIRECTIVES } from './lia-card.component';
