--- app/src/app/components/new-lia-card/index.ts
@@ -1,1 +1,0 @@
-export { LIA_CARD_DIRECTIVES } from './new-lia-card.component';
