+++ app/src/app/sections/home/home-page/home-page.component.html
@@ -1,16 +1,16 @@
 <div class="card-container">
-  <new-lia-card>
+  <lia-card>
     <card-title-group>
       <img card-img src="img/jobsity-icon.png" alt="Jobsity logo">
       <card-title>Live Interview Application</card-title>
       <card-subtitle>by Jobsity</card-subtitle>
     </card-title-group>
     <card-content>
-      <p>This is Home page for Live Interview Application</p>
+      <p>This is Home page for Live Interview Application project.</p>
       <p>In order to preview functionality in a test mode please click <strong>LAUNCH TEST LIA</strong> button.</p>
     </card-content>
     <card-actions>
       <button lia-button [routerLink]="['LiaCard', {userId: 1, liaId: 1}]">LAUNCH TEST LIA</button>
     </card-actions>
-  </new-lia-card>
+  </lia-card>
 </div>
