+++ app/src/app/sections/lia/lia-landing-page/lia-landing-page.component.css
@@ -5,6 +5,6 @@
   margin: 175px 0 185px 0;
 }
 
-new-lia-card {
+lia-card {
   width: 665px;
 }
