+++ app/src/app/sections/lia/lia-landing-page/lia-landing-page.component.html
@@ -1,5 +1,5 @@
 <div class="card-container" *ngIf="lia">
-  <new-lia-card>
+  <lia-card>
     <card-title-group>
       <img card-img src="img/jobsity-icon.png" alt="Jobsity logo">
       <card-title>Welcome to Jobsity Live Coding</card-title>
@@ -16,5 +16,5 @@
       <button lia-button (click)="launchLIA(lia)">Start Coding!</button>
       <a href="" lia-button-link>need help?</a>
     </card-actions>
-  </new-lia-card>
+  </lia-card>
 </div>
